+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 ## v4.66.0 Fixes
 
 - `[Fileupload Advanced]` Fix on max fileupload limit. ([#6625](https://github.com/infor-design/enterprise/issues/6625))
+- `[Datagrid]` Fixed redundant aria-describedby attributes at cells. ([#6530](https://github.com/infor-design/enterprise/issues/6530))
 
 ## v4.65.0